agents_v1/requirements.txt
# LangChain 1.0 - Core libraries
langchain>=1.0.0,<2.0.0
langchain-openai>=1.0.0
langchain-community>=0.4.0

# LangGraph 1.0 - State graph orchestration
langgraph>=1.0.0,<2.0.0

# Vector store for RAG agent
faiss-cpu>=1.8.0

# LLM providers
openai>=1.50.0

# Data validation
pydantic>=2.0.0,<3.0.0

# Environment configuration
python-dotenv>=1.0.0

# Real API integrations
requests>=2.31.0
numexpr>=2.8.0
duckduckgo-search>=7.0.0
yfinance>=0.2.0

# Observability - Phoenix Arize
arize-phoenix>=11.0.0
openinference-instrumentation-langchain>=0.1.0

# OPTIONAL: LangFuse (pip install langfuse)
# OPTIONAL: DeepEval (pip install deepeval)
# OPTIONAL: Guardrails AI (pip install guardrails-ai)
